Reified assertions.
Added assertion classes for coderefs, regular expressions and booleans.

Now Test::Unit::Assert::assert takes the first argument, the assertion,
creates an appropriate assertion object, and calls the object's do_assertion
method.

This seems to be a win.

Check out T::U::tests::SuiteTest and the 'basic_assertion' method. This is
possibly using a coderef for the sake of it, but it's one way of testing it
hard...

If you check out T::U::tests::TestTest you'll see a big win on the
regexp front.

Instead of

    $self->assert(scalar("foo" =~ /bar/), "Should not match");

we have

    $self->assert(qr/bar/, "foo");

(I haven't worked out how to name such an assertion, but the default
error message for that would look like:
    'foo' did not match /(?-xism:bar)/

which is pretty useful already)