Stopped Test::Unit::Assert::normalize_assert from dying when passed an object that can't 'do_assertion'. Instead treats it as the argument for Test::Unit::Assertion::Boolean::new.
(Came across this bug while testing some of my 'real' code).

This should also mean that, if you

    package Foo;

    use overload
        bool => sub {...}

Then assertion will continue to work properly.

I've added a test for the simple case --
T:U:t:TestTest::test_assert_with_non_assertion_object (essentially it's
always going to be true in this instance), but not for the case where
'bool' is overridden. I'll leave that as an exercise for the interested
reader (plus I'm not sure how to make inner classes do overrides).